# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Versa API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /growth/quote:
    post:
      operationId: previewBookingGrowthIncentive
      tags: [Growth]
      summary: Preview pricing without reserving a booking or incentive
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pitchId, startAt]
              properties:
                pitchId: { type: string, format: uuid }
                startAt: { type: string, format: date-time }
                promoCode: { type: string, maxLength: 32 }
      responses:
        "200":
          description: Non-reserving incentive quote
          content:
            application/json:
              schema:
                type: object
                required: [quote]
                properties:
                  quote: { $ref: "#/components/schemas/PricingQuote" }
        "400": { description: Invalid or inactive promotion }
        "404": { description: Pitch not found }
        "409": { description: Growth unavailable or redemption limit reached }
  /bookings/{bookingId}/quote:
    post:
      operationId: quoteBookingGrowthIncentive
      tags: [bookings]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: bookingId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                promoCode: { type: string, maxLength: 32 }
      responses:
        "200":
          description: Server-calculated incentive quote
          content:
            application/json:
              schema:
                type: object
                required: [quote]
                properties:
                  quote: { $ref: "#/components/schemas/PricingQuote" }
        "400": { description: Invalid or inactive promotion }
        "409": { description: Growth tools disabled or redemption limit reached }
  /owner/promotions:
    get:
      operationId: listOwnerPromotions
      tags: [bookings]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Owner promotions with redemption summaries
          content:
            application/json:
              schema:
                type: object
                required: [promotions]
                properties:
                  promotions:
                    type: array
                    items: { $ref: "#/components/schemas/GrowthPromotion" }
        "403": { description: Pro or Elite plan required }
    post:
      operationId: createOwnerPromotion
      tags: [bookings]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePromotionRequest" }
      responses:
        "201":
          description: Promotion created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GrowthPromotionResponse" }
        "403": { description: Pro or Elite plan required }
  /owner/promotions/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      operationId: updateOwnerPromotion
      tags: [bookings]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled: { type: boolean }
                code: { type: string }
                kind: { type: string, enum: [PERCENTAGE, FIXED] }
                value: { type: number, exclusiveMinimum: 0 }
                endsAt: { type: string, format: date-time, nullable: true }
                redemptionLimit: { type: integer, minimum: 1, nullable: true }
                perPlayerLimit: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Promotion updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GrowthPromotionResponse" }
    delete:
      operationId: deleteOwnerPromotion
      tags: [bookings]
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Unused promotion deleted }
        "409": { description: Promotion has redemption history; disable it instead }
  /owner/growth/analytics:
    get:
      operationId: getOwnerGrowthAnalytics
      tags: [bookings]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Scoped owner growth summaries
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GrowthAnalytics" }
  /owner/venues/{id}/streak:
    put:
      operationId: setVenueStreak
      tags: [venues]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        "200":
          description: Weekly streak setting updated
          content:
            application/json:
              schema:
                type: object
                required: [streak]
                properties:
                  streak:
                    type: object
                    required: [enabled, venueId]
                    properties:
                      enabled: { type: boolean }
                      venueId: { type: string, format: uuid }
        "403": { description: Pro or Elite plan required }
  /player/venues/{venueId}/streak:
    get:
      operationId: getPlayerVenueStreak
      tags: [venues]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: venueId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Player weekly streak progress and reward
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlayerVenueStreak" }
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GrowthPromotion:
      type: object
      required: [id, venueId, code, kind, value, enabled, perPlayerLimit, createdAt]
      properties:
        id: { type: string, format: uuid }
        venueId: { type: string, format: uuid }
        code: { type: string }
        kind: { type: string, enum: [PERCENTAGE, FIXED] }
        value: { type: string }
        enabled: { type: boolean }
        startsAt: { type: string, nullable: true }
        endsAt: { type: string, nullable: true }
        redemptionLimit: { type: integer, nullable: true }
        perPlayerLimit: { type: integer }
        reservedCount: { type: integer }
        redeemedCount: { type: integer }
        redeemedDiscountTotal: { type: string }
        createdAt: { type: string }
    GrowthPromotionResponse:
      type: object
      required: [promotion]
      properties:
        promotion: { $ref: "#/components/schemas/GrowthPromotion" }
    GrowthAnalytics:
      type: object
      required: [promotionCount, redeemedCount, discountTotal, activeStreakPlayers, nearRewardPlayers]
      properties:
        promotionCount: { type: integer }
        redeemedCount: { type: integer }
        discountTotal: { type: string }
        activeStreakPlayers: { type: integer }
        nearRewardPlayers: { type: integer }
    PlayerVenueStreak:
      type: object
      required: [enabled, progress, rewardAvailable, reward]
      properties:
        enabled: { type: boolean }
        rewardAvailable: { type: boolean }
        progress:
          type: object
          nullable: true
          required: [paidWeeks]
          properties:
            paidWeeks: { type: integer }
            expiresAt: { type: string, nullable: true }
        reward:
          type: object
          nullable: true
          properties:
            id: { type: string, format: uuid }
            status: { type: string, enum: [AVAILABLE] }
            expiresAt: { type: string, nullable: true }
    PricingQuote:
      type: object
      required: [currency, subtotal, discountAmount, payableAmount]
      properties:
        currency: { type: string, enum: [EUR] }
        subtotal: { type: string }
        discountAmount: { type: string }
        payableAmount: { type: string }
        incentiveType: { type: string, nullable: true, enum: [PROMOTION, STREAK_REWARD] }
        incentiveId: { type: string, format: uuid, nullable: true }
    CreatePromotionRequest:
      type: object
      required: [venueId, code, kind, value]
      properties:
        venueId: { type: string, format: uuid }
        code: { type: string, minLength: 3, maxLength: 32, pattern: '^[A-Z0-9][A-Z0-9_-]{2,31}$' }
        kind: { type: string, enum: [PERCENTAGE, FIXED] }
        value: { type: number, exclusiveMinimum: 0 }
        startsAt: { type: string, format: date-time, nullable: true }
        endsAt: { type: string, format: date-time, nullable: true }
        redemptionLimit: { type: integer, minimum: 1, nullable: true }
        perPlayerLimit: { type: integer, minimum: 1, default: 1 }
...